fix: bigedien and extract util to storeutils
