--- librevm/src/tests.rs
@@ -1,2 +1,0 @@
-#[test]
-pub fn test_create() {}
